Fixes the bug where it wasn't aiming correctly.
How I debugged it: I scaled health with the ideal turn to see what it was
trying to do, and changed the sprite frame depending on if it was turning
left or right. I ended up realizing that sign(ideal_turn) != sign(max_turn)
because max_turn was getting ideal_face's sign, not ideal_turn's!